requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
pydantic==1.10.7
numpy==1.24.3
opencv-python==4.7.0.72
pytesseract==0.3.10
Pillow==9.5.0
python-dateutil==2.8.2
SpeechRecognition==3.10.0
pyttsx3==2.90
